server/requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-configurations==2.4
django-cors-headers==3.13.0
django-js-asset==2.1.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
executing==1.2.0
flower==2.0.0
gunicorn==20.1.0
humanize==4.7.0
idna==3.4
importmagic3==0.2.0
inflection==0.5.1
ipdb==0.13.11
ipython==8.8.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jq==1.4.1
kombu==5.3.1
log-colorizer==2.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mercadopago==2.2.0
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prometheus-client==0.17.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7
redis==4.6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
spark-parser==1.8.9
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.8.1
typing_extensions==4.7.1
tzdata==2022.7
uncompyle6==3.9.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wdb==3.3.0
whitenoise==6.5.0
xdis==6.0.5
